Initial commit of conversion to service, monitor, and manager classes
